dev: upgrade tag board

+++ taglib/src/main/java/horizon/taglib/utils/Point.java
@@ -11,12 +11,12 @@ import lombok.Data;
  **/
 @Data
 public class Point {
-	private int x, y;
+	private double x, y;
 
 	public Point() {
 	}
 
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
